agent/requirements.txt
# LiveKit Agents with OpenAI and ElevenLabs plugins
# Requires Python >= 3.10, < 3.14
livekit-agents[openai,silero]~=1.0
livekit-plugins-elevenlabs~=1.3
python-dotenv>=1.0.0
httpx>=0.27.0
